update : refactoring
